+++ codex-rs/mcp-server/src/lib.rs
@@ -20,9 +20,10 @@ mod codex_tool_runner;
 mod exec_approval;
 mod json_to_toml;
 pub mod mcp_protocol;
-mod message_processor;
+pub(crate) mod message_processor;
 mod outgoing_message;
 mod patch_approval;
+pub(crate) mod tool_handlers;
 
 use crate::message_processor::MessageProcessor;
 use crate::outgoing_message::OutgoingMessage;
